kleinen fehler behoben. funktioniert aber immer noch nicht

+++ src/bootloader/stage3/headers/memory_management.h
@@ -19,7 +19,7 @@ uint8_t allocate_block(uint32_t *address, uint32_t *offset_addr, uint64_t size);
 void unblock_space(uint32_t *address, uint64_t size);
 void block_space(uint32_t *address, uint64_t size);
 void register_memspace(RAM_t *RAM);
-void memcp(uint32_t *src, uint32_t *dst, uint32_t start, uint32_t count, uint32_t offset_dst);
+void memcp(uint8_t *src, uint8_t *dst, uint32_t start, uint32_t count, uint32_t offset_dst);
 void memset(uint8_t *dst, uint32_t size, uint8_t value);
 
 // void output_mem_things();
